우선 테스트를 통과시킨다
