+++ app-server/subprojects/bounded_context/accessibility/application/src/main/kotlin/club/staircrusher/accessibility/application/port/in/result/RegisterBuildingAccessibilityResult.kt
@@ -7,7 +7,7 @@ import club.staircrusher.place.domain.model.Building
 
 data class RegisterBuildingAccessibilityResult(
     val building: Building,
-    val buildingAccessibility: BuildingAccessibility?,
+    val buildingAccessibility: BuildingAccessibility,
     val buildingAccessibilityComment: BuildingAccessibilityComment?,
     val accessibilityRegisterer: AccessibilityRegisterer?,
 )
